WIP: Assamble the initial game in Game
